+++ ExpandWorldPrefabs/HandleRPC.cs
@@ -29,7 +29,7 @@ public class HandleRPC
     {
       foreach (var state in states)
       {
-        if (state == "<always>" || requiredStates.Contains(state))
+        if (requiredStates.Contains(state))
           Handlers[hash] = handler;
       }
     }
@@ -76,15 +76,11 @@ public class HandleRPC
     if (health > 1E20) return false;
     if (health == wearNTear.m_health)
     {
-      var cancel1 = Manager.Handle(ActionType.State, ["repair"], zdo);
-      var cancel2 = Manager.Handle(ActionType.Repair, [], zdo);
-      return cancel1 || cancel2;
+      return Manager.Handle(ActionType.State, ["repair"], zdo);
     }
     else
     {
-      var cancel1 = Manager.Handle(ActionType.State, ["damage"], zdo);
-      var cancel2 = Manager.Handle(ActionType.Damage, [], zdo);
-      return cancel1 || cancel2;
+      return Manager.Handle(ActionType.State, ["damage"], zdo);
     }
   }
 
@@ -97,9 +93,7 @@ public class HandleRPC
     if (pars.Length < 2) return false;
     var trigger = (string)pars[1];
 
-    var cancel1 = Manager.Handle(ActionType.State, ["action", trigger], zdo);
-    var cancel2 = Manager.Handle(ActionType.State, [trigger], zdo);
-    return cancel1 || cancel2;
+    return Manager.Handle(ActionType.State, ["action", trigger], zdo);
   }
   static readonly int SetTargetHash = "RPC_SetTarget".GetStableHashCode();
   static readonly ParameterInfo[] SetTargetPars = AccessTools.Method(typeof(Turret), nameof(Turret.RPC_SetTarget)).GetParameters();
@@ -114,18 +108,15 @@ public class HandleRPC
     if (targetZDO == null) return false;
     var targetPrefab = ZNetScene.instance.GetPrefab(targetZDO.GetPrefab());
     if (!targetPrefab) return false;
-    var cancel1 = Manager.Handle(ActionType.State, [targetPrefab.name], zdo);
-    var cancel2 = Manager.Handle(ActionType.State, ["targeting", targetPrefab.name], zdo);
-    var cancel3 = Manager.Handle(ActionType.State, ["target"], targetZDO);
-    return cancel1 || cancel2 || cancel3;
+    var cancel1 = Manager.Handle(ActionType.State, ["targeting", targetPrefab.name], zdo);
+    var cancel2 = Manager.Handle(ActionType.State, ["target"], targetZDO);
+    return cancel1 || cancel2;
   }
   static readonly int ShakeHash = "RPC_Shake".GetStableHashCode();
   static readonly ParameterInfo[] ShakePars = AccessTools.Method(typeof(TreeBase), nameof(TreeBase.RPC_Shake)).GetParameters();
   private static bool Shake(ZDO zdo, ZRoutedRpc.RoutedRPCData data)
   {
-    var cancel1 = Manager.Handle(ActionType.Damage, [], zdo);
-    var cancel2 = Manager.Handle(ActionType.State, ["damage"], zdo);
-    return cancel1 || cancel2;
+    return Manager.Handle(ActionType.State, ["damage"], zdo);
   }
   static readonly int OnStateChangedHash = "RPC_OnStateChanged".GetStableHashCode();
   static readonly ParameterInfo[] OnStateChangedPars = AccessTools.Method(typeof(Trap), nameof(Trap.RPC_OnStateChanged)).GetParameters();
@@ -136,9 +127,7 @@ public class HandleRPC
     if (pars.Length < 2) return false;
     var state = (int)pars[1];
     if (state == 0) return false;
-    var cancel1 = Manager.Handle(ActionType.State, [state.ToString()], zdo);
-    var cancel2 = Manager.Handle(ActionType.State, ["trap", state.ToString()], zdo);
-    return cancel1 || cancel2;
+    return Manager.Handle(ActionType.State, ["trap", state.ToString()], zdo);
   }
   static readonly int SetSaddleHash = "SetSaddle".GetStableHashCode();
   static readonly ParameterInfo[] SetSaddlePars = AccessTools.Method(typeof(Tameable), nameof(Tameable.RPC_SetSaddle)).GetParameters();
@@ -223,9 +212,7 @@ public class HandleRPC
     if (pars.Length < 3) return false;
     var index = (int)pars[1];
     var health = (float)pars[2];
-    var cancel1 = Manager.Handle(ActionType.Damage, [], zdo);
-    var cancel2 = Manager.Handle(ActionType.State, ["damage", index.ToString(), Helper.Format2(health)], zdo);
-    return cancel1 || cancel2;
+    return Manager.Handle(ActionType.State, ["damage", index.ToString(), Helper.Format2(health)], zdo);
   }
   static readonly int HideHash = "Hide".GetStableHashCode();
   static readonly ParameterInfo[] HidePars = AccessTools.Method(typeof(MineRock), nameof(MineRock.RPC_Hide)).GetParameters();
@@ -235,9 +222,7 @@ public class HandleRPC
     data.m_parameters.SetPos(0);
     if (pars.Length < 2) return false;
     var index = (int)pars[1];
-    var cancel1 = Manager.Handle(ActionType.Damage, [index.ToString()], zdo);
-    var cancel2 = Manager.Handle(ActionType.State, ["damage", index.ToString()], zdo);
-    return cancel1 || cancel2;
+    return Manager.Handle(ActionType.State, ["damage", index.ToString()], zdo);
   }
   static readonly int SetVisualItemHash = "SetVisualItem".GetStableHashCode();
   static readonly ParameterInfo[] ItemStandPars = AccessTools.Method(typeof(ItemStand), nameof(ItemStand.RPC_SetVisualItem)).GetParameters();
@@ -254,9 +239,7 @@ public class HandleRPC
     var quality = (int)pars[3];
     var state = item == "" ? "<none>" : item;
 
-    var cancel1 = Manager.Handle(ActionType.State, [state, variant.ToString(), quality.ToString()], zdo);
-    var cancel2 = Manager.Handle(ActionType.State, ["item", state, variant.ToString(), quality.ToString()], zdo);
-    return cancel1 || cancel2;
+    return Manager.Handle(ActionType.State, ["item", state, variant.ToString(), quality.ToString()], zdo);
   }
   static readonly int SetArmorVisualItemHash = "RPC_SetVisualItem".GetStableHashCode();
   static readonly ParameterInfo[] ArmorStandPars = AccessTools.Method(typeof(ArmorStand), nameof(ArmorStand.RPC_SetVisualItem)).GetParameters();
@@ -272,9 +255,7 @@ public class HandleRPC
     var variant = (int)pars[3];
     var state = item == "" ? "<none>" : item;
 
-    var cancel1 = Manager.Handle(ActionType.State, [state, variant.ToString(), slot.ToString()], zdo);
-    var cancel2 = Manager.Handle(ActionType.State, ["item", state, variant.ToString(), slot.ToString()], zdo);
-    return cancel1 || cancel2;
+    return Manager.Handle(ActionType.State, ["item", state, variant.ToString(), slot.ToString()], zdo);
   }
   static readonly int AnimateLeverHash = "RPC_AnimateLever".GetStableHashCode();
   static readonly ParameterInfo[] AnimateLeverPars = AccessTools.Method(typeof(Incinerator), nameof(Incinerator.RPC_AnimateLever)).GetParameters();
@@ -299,9 +280,7 @@ public class HandleRPC
     var item = (string)pars[2];
     var state = item == "" ? "<none>" : item;
 
-    var cancel1 = Manager.Handle(ActionType.State, [slot.ToString(), state], zdo);
-    var cancel2 = Manager.Handle(ActionType.State, ["item", slot.ToString(), state], zdo);
-    return cancel1 || cancel2;
+    return Manager.Handle(ActionType.State, ["item", slot.ToString(), state], zdo);
   }
 
 
@@ -462,10 +441,10 @@ public class HandleRPC
   // For some RPCs this is always the player what did the action.
   private static readonly (int Hash, RPCHandler Handler, string[] States)[] AllAvailableHandlers = [
     (RepairHash, WNTHealthChanged, ["damage", "repair"]),
-    (SetTriggerHash, SetTrigger, ["<always>", "action"]),
-    (SetTargetHash, SetTarget, ["<always>", "target", "targeting"]),
+    (SetTriggerHash, SetTrigger, ["action"]),
+    (SetTargetHash, SetTarget, ["target", "targeting"]),
     (ShakeHash, Shake, ["damage"]),
-    (OnStateChangedHash, OnStateChanged, ["<always>", "trap"]),
+    (OnStateChangedHash, OnStateChanged, ["trap"]),
     (SetSaddleHash, SetSaddle, ["saddle", "unsaddle"]),
     (SayHash, Say, ["say"]), // Uses ActionType.Say
     (FlashShieldHash, FlashShield, ["flash"]),
@@ -474,11 +453,11 @@ public class HandleRPC
     (WakeupHash, Wakeup, ["wakeup"]),
     (SetAreaHealthHash, SetAreaHealth, ["damage"]),
     (HideHash, Hide, ["damage"]),
-    (SetVisualItemHash, SetVisualItem, ["<always>", "item"]),
+    (SetVisualItemHash, SetVisualItem, ["item"]),
     (AnimateLeverHash, AnimateLever, ["start"]),
     (AnimateLeverReturnHash, AnimateLeverReturn, ["end"]),
-    (SetArmorVisualItemHash, SetArmorVisualItem, ["<always>", "item"]),
-    (SetSlotVisualHash, SetSlotVisual, ["<always>", "item"]),
+    (SetArmorVisualItemHash, SetArmorVisualItem, ["item"]),
+    (SetSlotVisualHash, SetSlotVisual, ["item"]),
     (MakePieceHash, MakePiece, ["piece"]),
     (OnEatHash, OnEat, ["eat"]),
     (OnDeathHash, OnDeath, ["death"]),